Refactor 4 move kafka function to kafka file
